+++ methods/src/commonMain/kotlin/ru/krindra/vkkt/methods/Gifts.kt
@@ -17,7 +17,7 @@ class Gifts(
      * @param count Number of gifts to return.
      * @param offset Offset needed to return a specific subset of results.
      */
-    suspend fun get(userId: Long? = null, count: Int? = null, offset: Int? = null): GiftsGetResponse {
+    suspend fun get(userId: Int? = null, count: Int? = null, offset: Int? = null): GiftsGetResponse {
         val response = method("gifts.get", mapOf("user_id" to userId, "count" to count, "offset" to offset))
         return decodeResponse(response, json)
     }
